Refactor: Use dataService for issue linking and validation
