+++ src/data/inMemoryStorage.ts
@@ -33,25 +33,3 @@ export interface Webhook {
 export const issues: Issue[] = [];
 export const boards: Board[] = [];
 export const webhooks: Webhook[] = [];
-
-export const addIssueLink = (inwardKey: string, outwardKey: string) => {
-  const inwardIssue = issues.find(issue => issue.key === inwardKey);
-  const outwardIssue = issues.find(issue => issue.key === outwardKey);
-
-  if (inwardIssue) {
-    if (!inwardIssue.linkedIssues) {
-      inwardIssue.linkedIssues = [];
-    }
-    if (!inwardIssue.linkedIssues.includes(outwardKey)) {
-      inwardIssue.linkedIssues.push(outwardKey);
-    }
-  }
-  if (outwardIssue) {
-    if (!outwardIssue.linkedIssues) {
-      outwardIssue.linkedIssues = [];
-    }
-    if (!outwardIssue.linkedIssues.includes(inwardKey)) {
-      outwardIssue.linkedIssues.push(inwardKey);
-    }
-  }
-};
